+++ rustfst/src/fst_traits/final_states_iterator.rs
@@ -3,42 +3,42 @@ use crate::semirings::Semiring;
 use crate::StateId;
 
 #[derive(Debug)]
-pub struct FinalState<W: Semiring> {
+pub struct FinalState<'f, W: Semiring> {
     pub state_id: StateId,
-    pub final_weight: W,
+    pub final_weight: &'f W,
 }
 
 /// Trait to iterate over the final states of a wFST.
-pub trait FinalStatesIterator<'a> {
-    type W: Semiring;
-    type Iter: Iterator<Item = FinalState<Self::W>>;
-    fn final_states_iter(&'a self) -> Self::Iter;
+pub trait FinalStatesIterator<'f> {
+    type W: Semiring + 'f;
+    type Iter: Iterator<Item = FinalState<'f, Self::W>>;
+    fn final_states_iter(&'f self) -> Self::Iter;
 }
 
-impl<'a, F> FinalStatesIterator<'a> for F
+impl<'f, F> FinalStatesIterator<'f> for F
 where
-    F: 'a + Fst,
+    F: 'f + Fst,
 {
     type W = F::W;
-    type Iter = StructFinalStatesIterator<'a, F>;
-    fn final_states_iter(&'a self) -> Self::Iter {
+    type Iter = StructFinalStatesIterator<'f, F>;
+    fn final_states_iter(&'f self) -> Self::Iter {
         StructFinalStatesIterator::new(&self)
     }
 }
 
-pub struct StructFinalStatesIterator<'a, F>
+pub struct StructFinalStatesIterator<'f, F>
 where
-    F: 'a + Fst,
+    F: 'f + Fst,
 {
-    fst: &'a F,
-    it: <F as StateIterator<'a>>::Iter,
+    fst: &'f F,
+    it: <F as StateIterator<'f>>::Iter,
 }
 
-impl<'a, F> StructFinalStatesIterator<'a, F>
+impl<'f, F> StructFinalStatesIterator<'f, F>
 where
-    F: 'a + Fst,
+    F: 'f + Fst,
 {
-    fn new(fst: &'a F) -> StructFinalStatesIterator<F> {
+    fn new(fst: &'f F) -> StructFinalStatesIterator<F> {
         StructFinalStatesIterator {
             fst,
             it: fst.states_iter(),
@@ -46,19 +46,18 @@ where
     }
 }
 
-impl<'a, F> Iterator for StructFinalStatesIterator<'a, F>
+impl<'f, F> Iterator for StructFinalStatesIterator<'f, F>
 where
-    F: 'a + Fst,
+    F: 'f + Fst,
 {
-    type Item = FinalState<F::W>;
+    type Item = FinalState<'f, F::W>;
 
     fn next(&mut self) -> Option<Self::Item> {
         while let Some(state_id) = self.it.next() {
-            if let Some(final_weight) = self.fst.final_weight(state_id) {
+            if let Some(final_weight) = unsafe { self.fst.final_weight_unchecked(state_id) } {
                 return Some(FinalState {
                     state_id,
-                    // TODO: No need to clone here.
-                    final_weight: final_weight.clone(),
+                    final_weight: &final_weight,
                 });
             }
         }